+++ R/clean_parameters.R
@@ -70,6 +70,10 @@ clean_parameters.default <- function(x, group = "", ...) {
       "scale"
     } else if (grepl("marginal", i, fixed = TRUE)) {
       "marginal"
+    } else if (grepl("intercept", i, fixed = TRUE)) {
+      "intercept"
+    } else if (grepl("correlation", i, fixed = TRUE)) {
+      "correlation"
     } else {
       "conditional"
     }
@@ -188,6 +192,29 @@ clean_parameters.wbm <- function(x, ...) {
 
 #' @export
 clean_parameters.wbgee <- clean_parameters.wbm
+
+
+
+#' @export
+clean_parameters.glmm <- function(x, ...) {
+  pars <- find_parameters(x, effects = "all", component = "all", flatten = FALSE)
+
+  l <- lapply(names(pars), function(i) {
+    data.frame(
+      Parameter = pars[[i]],
+      Effects = i,
+      Component = "",
+      Group = "",
+      Function = "",
+      Cleaned_Parameter = clean_names(pars[[i]]),
+      stringsAsFactors = FALSE,
+      row.names = NULL
+    )
+  })
+
+  out <- .remove_backticks_from_parameter_names(do.call(rbind, l))
+  .remove_empty_columns_from_pars(out)
+}